+++ packages/expo-file-system/ios/Next/FileSystemPath.swift
@@ -12,15 +12,15 @@ internal class FileSystemPath: SharedObject {
   func validatePermission(_ flag: EXFileSystemPermissionFlags) throws {
     try ensurePathPermission(appContext, path: url.path, flag: flag)
   }
-  
+
   func validateCanCreate(_ options: CreateOptions) throws {
-    if(!options.overwrite) {
+    if !options.overwrite {
       guard !(try exists) else {
         throw UnableToCreateException("it already exists")
       }
     }
   }
-  
+
   var exists: Bool {
     get throws {
       FileManager.default.fileExists(atPath: url.path)